fix wbtc and replace
