--- projects/wbtc/index.js
@@ -1,410 +1,0 @@
-const sdk = require("@defillama/sdk");
-const { sumTokensExport } = require("../helper/sumTokens");
-const ADDRESSES = require('./helper/coreAssets.json')
-
-
-// WALLETS FROM HERE https://wbtc.network/dashboard/audit
-const owners = [
-  "31qyn946XELbPKMaSWhk2SfFN8ikqLJ1GY",
-  "31rYKY9PysQgRoc6eB9fwJ3e89F5dcvRM4",
-  "32Ks4Rtb1A5NtzexGq6zEeqHta6hoqvCZW",
-  "32hJnfPp6ZDnmkJ3tosoo6PvEemFr1z1Bd",
-  "32hPQh4EMBS14BMutrW2WJ4V76yxuXHJdF",
-  "32iT1QaoQD61rZgCNLccNqAdbnx4HwgvZE",
-  "32m8wzKabRRF7o6JRXN1srLSA5FzND769o",
-  "32oLNY3suWAZvYTBeAaFFSk32rJtS8w9i8",
-  "32okkzumvdoocdjN18SzbnTTNd2omQvvt3",
-  "337ot6fHGs4NCZwg2cKwShhrYX46g8ACtQ",
-  "33C6kjmmGp2mrS6coGn627an3xpiRpEyhX",
-  "33DM5s2vPgX2P52DpurbXdVYa8dGwDCsJ3",
-  "33K5r4QntjmmD4mzSHq3WQJQtf4Dw8jvcZ",
-  "33ZA8VjwL2oucQceCP9WLEVunLkLvhaaPb",
-  "33epzdB4nk3FWHtUpHMQrzTWxiehSti11v",
-  "33jqiX8NtPd8empn7ibEc2tn25GUx8iUE8",
-  "33xk2gq91LNUaip9bPnoECtfaA1RjAcrSo",
-  "33xpP9NhBBmQt3XS1mD2g4pi1FZNuUFVR8",
-  "343uVxqsriiveWxekC6s3mNaERzsbfke3h",
-  "34MSicAL7qVGkevFPLwyc9KohGzoUSnu3Q",
-  "34whRPeeSYTPcihybfXmkG2uoShUdT3d2d",
-  "35AY8sKs3yt3FbZpsmAtgvSXBZtvqqwc8e",
-  "35HfRwMwqB7sNTwijoG4fGHBmfWMm48aaB",
-  "35QASqiGmQmHoccRc8pk9EtPVDFZDrFA59",
-  "35SZSm1CDcz7p27Hj78xZACv4Ak45EKGpf",
-  "35ULMyVnFoYaPaMxwHTRmaGdABpAThM4QR",
-  "35XnFgLYBuYogTx9HHGbJCRGfyqyC9ZtDG",
-  "35ZmRhYBwDfvoASsw35KNJxwypvsX9xD14",
-  "35sfxCN1Gz37dq9KkpTfW56W59CK46JMY6",
-  "36WQ2FEUnTmt67xXNZgQMSFQZjK6AJxLCZ",
-  "36hbDz8GSBpQdEf2UzcKieCNxSA3HS17JX",
-  "36jYq75ewM9AMTfnKRTwkPBBWwJtGg4CsE",
-  "371SbkDHpTvRuYTwayWsBNsXK93Tj4sDAd",
-  "37Tgw6PStVEAyfAJT8yEvLAcxkoRYK3hDz",
-  "37Tob2tV9LRVzJng6Z4YCLyPsmq8n2xMQi",
-  "37aZoS67WApGWTDsn2rYDsxebhTdZSR1sh",
-  "37hV1tJYctQLPin7vCSjpDL34PpiiUTyTx",
-  "38J1ce6tPUgy958VBkGGa2AmFU3LQLJ2Gv",
-  "38MHxNgiHENWExwCe3EEWYmfvgFM3i1WVW",
-  "38R6URq21qncKbYiiRXPi2MVx8xYRgDtWD",
-  "38TQEEGTmmPa1rghsYGwZhapiDQsSEnkub",
-  "38b9cazfu4SxqoahgdiHEMSo1aNDdKquRm",
-  "38bzm6nhQMFJe71jJw1U7CbgNrVNpkonZF",
-  "38rK7BfeiegdchggPCk2CZc3fuy8ozwzPP",
-  "39RMkGSQUQjifV2TpgRNYW73JK38xwA3NJ",
-  "39WQqCosC8ZD4S9XBPHRnnRUeVRvNctnnm",
-  "39e7mxbeNmRRnjfy1qkphv1TiMcztZ8VuE",
-  "39mayj82aZdb6UWs9qQAZZrC3VzfxUQP7c",
-  "39pEDUwFFj28sN5muXrmT5mAbDJSf6qEL2",
-  "3A2wJ7LBiR4NhPbJSLNVfzZk6vDeYYPrit",
-  "3ANaBZ6odMrzdg9xifgRNxAUFUxnReesws",
-  "3B9UDWZNPdtL81JLnnPEJoiurcmaGx8DZm",
-  "3BJJ68V6H1CoEx4MWdhqVk2kFpPUNVbs3T",
-  "3BTqdvZt5MYYWLAGXt8psAMSAa8Gh2AVPA",
-  "3BY62gfDUmuzMp1T9Wx4LnMRk3oNbQsKyQ",
-  "3Brhg9KYdB9oDrUtvKJ9k7mbn7zzHnMV51",
-  "3DDUy5Ze4TaqKFpktAhHcmmffz9KKGtTrb",
-  "3DKc1cFmTC4BJFcoyeN7fDp1oavTM4SvuX",
-  "3DeAGDRtQ984GazpLqfpS4kRW8NXBei5V8",
-  "3DiJgisVit4pqj4ByKxi5p1qdjGDxpGfR4",
-  "3ERNo3RAgP8o5jqELPKUEV47ADm4JtvKBd",
-  "3Epe1CixppBP9QSQ7DkvsWNKsgifTaxMY2",
-  "3ErNFK17MbH3GAGfakMwTrKo8uHmBuuVX1",
-  "3FFgKaYkEf1M73QtzuY9DGqC7VeM2sAQhT",
-  "3G3kGKAD5BgWZu91ejMXsW5LxEMhKZj46Z",
-  "3GbZUJ2oaUG3rZTvAwTCqpsNMWGm6j62r1",
-  "3GnY2SF2DoWRwWyQb6jThRJ5xYzNWFfd7J",
-  "3Go9rkgYQ3oLXfbbiDFENXmWZXQDaMjvtx",
-  "3GohY4RNGXjRH5KhByj5m4Amxq9jTsGfBr",
-  "3HFrW63zQ58PypYhSYgHB2gij5VirsyRAm",
-  "3HLxZqoEHdGPV9oFZQxVfG41z1uP1LrNaJ",
-  "3HVCDPPbAqiRwRjXhvu7PSDkLGgEajexuH",
-  "3Hdj2n5DEKPGsRd5md85exAXdwnRbomWTy",
-  "3He1eqsBZM5kbtZGuUDX7ysfEVWCjweZtJ",
-  "3JDewabYE785vTmaQe3MuHvUvEZJQTx8HU",
-  "3JJpCZCk4h4TpQeU7SA1yhH768Xgbtdbfg",
-  "3JTWYpPz7JkvhnAN7XKzYK31AfmxEUzRb8",
-  "3JbEPhuab1Gx6yrY8dD6wF1NrVAn3eG8CD",
-  "3JcNkH5QWYuSfiQYADq4yV2hAYnQ8KC74t",
-  "3JfZb6DbsRaQExeHAYmFa3s3azckqdyQr5",
-  "3K2mpXMDFe7cT1baogitkRGeXqCYeDw6J5",
-  "3KT7jvZyvhok3fkZt8Y9w8tcqpyQW5Jg8s",
-  "3KTeq879YjzhqkAXzZmdapJAVC6qz5qEth",
-  "3KUrPEGy9VFeQHZYJSGZpE78X5zf1n48pP",
-  "3KXKTXR9YxHuhm863nqnSrKznkYtcx7wC7",
-  "3KkkHYZywtjNy2eoVvL1xLo2Mnzz3gaDqQ",
-  "3KwPoNu1FxihhyRKjv1VG6C3FPy8vgjyse",
-  "3LScoRmRA3kknxF5dLxz6wfw3z4LXRiVyT",
-  "3LYNPfnfmGRQcheXUZFRNEYPDDmLvDSgAE",
-  "3LoYYxQis416ykPKGNuWe4iLdfSPWCHjgd",
-  "3LrxWneLHwvbNrMKXorkc17wxHqCCpNGkq",
-  "3MJ4rpTvuiTxYhTppNx7NokNUDZg26ofi1",
-  "3MPcUSHeWCSxCUsJtTS2L2zZfWLvorNFef",
-  "3MUJjyx13waUriNPkVnemdoJAiJT6nBmef",
-  "3Mzdh31vo6wUPTWnZ7GncPPWXA1WDopkhm",
-  "3NNsvp7bfeihJTiGbiL3t6jH3jUbRERsB7",
-  "3NsC2NXzKum7sVWQLQ3YYB3o9JbTz5FoaV",
-  "3PfcB36gDa47YkrbRtzZSwmrKSMsBSrhr6",
-  "3PggDrcp4et9phMamnSenN8HeNPP2YCuCX",
-  "3QB2qhnj5Xxwhh3GKcfwQDSkL91BCCn5cp",
-  "3QSd3JjJ35VAf5Shv3EPVpGCi5eVMQr6dk",
-  "3QWaxXXXbobqpWprekwzANeFuNrYXmaNXU",
-  "3QjqS6A7uyoCx1vcXQoqtLatubFmsHe1ps",
-  "bc1q02hdwthfslrempcvrgnrelsluszmt20qjaahg3tn2plse5pc3evsskz9g6",
-  "bc1q07jp79h30qmh39pkdmntaycwezmfl2njzvp85z8sdthhp2cus42qrwzglw",
-  "bc1q094s5w7semhdkcgzxrhxe6am449prjhsend8p83fhk9jv6rnr0eq6af0pp",
-  "bc1q0asc7qsj90esjzfmgz7r79tc5vd455yugwx0q4eqdqhqs82t0s2sfhtec2",
-  "bc1q0de5mlkcvundsnzdsmj6padctjh25jgsq3tcyu6d2qcgjnz0h8us2sws8a",
-  "bc1q0nnem9u53fcxcnmuutqw7j458hn56ljte34edvstjfyz8rfhxyesju3hhx",
-  "bc1q0s0enxvmfvfg3dj0rza2el0vz7p3a6xk2juv7vztw22l4g35rknsmv58se",
-  "bc1q0tckafgtukhn78d32ljl8c3hr0an5jq08599gf08mqxeqlahdg3s5wvuz3",
-  "bc1q0tvp3naff0auk3p4q99k3s9ltlf2w97uclrjspf3lvjkjywsj8lqvdx46f",
-  "bc1q0vthrmhktt2mhew5rq6e5q3qlj573uxcfqry5h9ea3gh4l9fz6hq999wz3",
-  "bc1q20mxuuv6ta5uy6m3s34lzwgs6p4qme4wyvv9y2cawd8a45y84susw2c9rk",
-  "bc1q2420wgxmqkxjkpkkdd9hxflxuqxhuqnt7atg0w3q3ypv0qfel0yqtanfuy",
-  "bc1q28pdpztsjuy4lqsa5hs249vq8ftykkt06aa8w97hgjxuvqz30trqqz5jdh",
-  "bc1q2j0ks6v6swlavvg20adsdn7hf9avm0979jc003g9lkg9dx0trz3sma7q62",
-  "bc1q2njvavlw9esncygac77d3jxjfwu877f23gxa88x33vjf3xa24lqq0xxetl",
-  "bc1q2nra9wwwq7pu4cg8zwycp8dzg9hcv8ywlj356p4ahjy7l80khphqvjdyfy",
-  "bc1q2qux9ad0egep5v5f2sup5hyszrl3nrfzyegmnmccmw7yk9chplhsy5sn4l",
-  "bc1q2wd3mfdcvjrxlccpl7wrzunzch7xaymtfzuvvx2tld760j3x5chqgsget8",
-  "bc1q33qexf56q60lxfpjhvgkj0l0lcgsapa7lsugkcwnyrceuw7kd89qsmy8h7",
-  "bc1q349jnpnefh952da4unfapsd3a6zmhgw8ctd2nk4ut40r37ax0d2qn0dluu",
-  "bc1q3d9jgd2yfx49986hjek9ulcj55umf6uyssrxqdmma7w7f7cyhc7qqyqvzr",
-  "bc1q3dyv677fn6feur2qjaqjdcftzgrmrwa7mjnvq7y8wxtka4ceu8tswfcvwm",
-  "bc1q42vvtdvdkdzj7glr6mtd52vgxddagktxusha6dxl2fes2dernqwqzff9k3",
-  "bc1q46uyezygewcl972rkp6z4pdtay669pxwrl8q70mgnda4cyf6yz0qj6rcsn",
-  "bc1q47z7kxma7lp7chnkaucee7ey9qds292kylz6rqvdlusu839ved9stns32j",
-  "bc1q48j4p67jmgpdpzxtu5tv3rendwvf6jjtqs2gt8ed96syxuqrsdfsl788un",
-  "bc1q4fpz2ntr5zlz6jga924hmaqc4a86s8luvfym3ff0qqr6gxykp60qeqqdn7",
-  "bc1q4gs9f4nhla3ufe6sumn8k0smcc5vxr0524nz3d5nhww70dr703lse46rmp",
-  "bc1q4gvftsdjy8wd6ga8c3a66rg4q6lxqvgak6edgnz7whs690vy2x8st6at64",
-  "bc1q4tl9hjvv3yvvsgqdajd492etlg8phrtml36ee5n0dt3mylmce23qnd5fdm",
-  "bc1q4zvuejph0q3z2c0narhgjtz6w47h6rst5eyyvr270lxh4tfqzluqrx4h4f",
-  "bc1q5a0t7av9xtay7av70fdwscsnx62rssut3mg7dclwmmsj0nxq5z3qgm8t49",
-  "bc1q5j8dyy5u94j0ulfg98l6jwsgjfq7dna0wwcwqvjcmwkgkwtcf30sm856qg",
-  "bc1q5ltp433ew9q98antza5s68h46g404eeaj6jkg4u6799q9dmkdngsmh6zny",
-  "bc1q5mq8f6rh85jwk0agf8huhtjp6syxcu7q4dk7alda9c2cedflj8js0u8j9f",
-  "bc1q5s4sa6s4vgst2pt97gvnxrmh9z8xa2j7kj370400fpx06pkk9uls7vzaaq",
-  "bc1q5sn9gqy6hd46yw545946f6a5crqryskhs9gge5z3gkp5d9luezmqllv7wz",
-  "bc1q5wj7yc27dzvnejeq9h543d6xt4cd90q8v8569tpzm6g5k3k252wq9jwuza",
-  "bc1q5xfafyynujpe4xm3sau92sy8rar3auf80vz3as6cjrn7pyvncfxq34hhcv",
-  "bc1q63yjt8zpr2kcegcmjr3vyqukm9zhlr432mcgss027lzs36qeh7qse8ct8w",
-  "bc1q67r0cfzcaqqm70sr00jq9fdyaplwcg24rtskylppej2qpgscjw7q5nrwf9",
-  "bc1q68mfyj37aurdzxwpq93vka6qkvhmcjl02l66va6vy8yktegp6nvs5knqe2",
-  "bc1q6e05t64zsdffvjxgw9njyxt5yhgq7xt2xxsufxk5npfcz4aewrgsyccu28",
-  "bc1q6hcsftu0hmm7q67cec272y8sxwhkq0pnpymshuty9x8wmymmjhdsquhzh4",
-  "bc1q6j7yxe948zd35xsrx3jucflxeq65unmrwusrm9a6c82lj459zmyskn5xk3",
-  "bc1q6kumts4ps94pyp8mfdn7yyy9p7y8n0wvgq760xpyk2r005rsh0dqdxxp22",
-  "bc1q6lmpnckv0zkdkv6xkvd2m2uy8mxetnz9g3qxkp3kpkvfhnsv0w7slkje25",
-  "bc1q6vk4rdxdnuklg5m4ltm866v6scz6lyuxdhj50qp6p4en8vtt23js9wpzgj",
-  "bc1q6vs72ztq8xchew26exy4kdnjxzrlxz43ny4lfckg95z5ka6ge6rs4rc5g2",
-  "bc1q6xr3ys66ry7jsfdfrlz82kfkwwnmdqtjh0ctm7y974kvufrwwezs6clrgr",
-  "bc1q6yk7s5g8u2ev7k9sw4zdazlfxj4vtduxgpekxvttzhju00c0yshs7ymvzl",
-  "bc1q6yzjmcudjlvq5nc0un27kfzwr283u0wdm53kqhelsja428t8jjfqpxc32j",
-  "bc1q76gspqpw24cqsptydstwy3egft798n2zq39gxfmfuxd5fvhh8p9s2s7c7f",
-  "bc1q78df3e2evtq70ntdcmcwzn0gs8a9zrlzh3p3xfrrt7p7s589pqesccuz22",
-  "bc1q78tqk9mn587sgnqqthsyrawmwpne7ssl4a4wlyasq4c9jqjtndwsyj22kn",
-  "bc1q79crmsa0k8g8k735nqqlctxqhz6eeugw2qv3x0sskdvsmsvjq7rsujhrr9",
-  "bc1q7d5q7e53k4w3c6n28tjrdvmlxl8skxutykk6fs9fgh79h0k9xc4sr9hzuy",
-  "bc1q7du4789yhw996lqwx6scre99sfzy2e53ax9502daxscsye7vvmksy8r743",
-  "bc1q7eaej94t49psde7mkmy6dd7vgn8f0zg45yqalpysy7u47y3z00lqg3x7fn",
-  "bc1q7l6c5ddseh5h54680ae0ufkx29y6rnl5yr8vhr6nfwc5smk8p0ysexsgsw",
-  "bc1q7pnj8k3tphfvcq7vrzlja5vqdqp7xed9rfn9hdl4ehgq09la9yhs5dpurz",
-  "bc1q7ps737rkza38ha5ysg3k4sjrd2xauexthkw23lxt74lpr70gum4s2jp8uk",
-  "bc1q7ta8kheq8ellkn6fmmkzxj47vq0uqcd53zd8n7g45hl945d6eals3d4ljj",
-  "bc1q7wcl3plp72cghpd6lapucmxtaa5fq4w4wmhswu6xjqhrmt9jsh7sfapsw4",
-  "bc1q7zdunrqm74erprv46nyryjkvl5nuymgmn2ek3x7mgck48umga5mstxdl9p",
-  "bc1q8fkquy9luszzu2mlsad3fsjgx433vp4qx69wswz5skhz65txv2lssgaua9",
-  "bc1q8mfdkttc6n804j43jmcqkftdug7yje77a55qeyhsy5ej7ca65wgscrvyd2",
-  "bc1q8nm7r22vw6lnmzl52lqvnlrft6hm36ye0087e984feavjmej4spq8y57z9",
-  "bc1q8pqeaj8c647ascnjpc582dzwfw8xstvyderzxzvc2dk09j8d4rasvctr5f",
-  "bc1q8uldtn4rnr23djmpvgwacs6tcazepkflhmej9t5r3afc9lldq4mqrl6zna",
-  "bc1q93rsljth5uxrpye4na8688r043lye0ny7uyexr4n24hf9wlgdmns9zcjn5",
-  "bc1q95ptqclhvj9vrsfnexfd4fs3l95dyyrt3cq3qyfpgxsmfuhspj3qq5gcqt",
-  "bc1q9988y3xw3mhv0tnrugu8c2vsw7e6f7ywrdaya6wcrcertw8rz54snxgnl9",
-  "bc1q9cs23smvx9c4q6eqcjpxgp0yrw6ar62w2x4mwe4dd88hqsfg35yq2pfk3y",
-  "bc1q9d6ayjfl3jagylp7x86v3dzzk65u0gt34mw74w9fxh0nyaf24slqp6sat4",
-  "bc1q9ehldvjkmd3s28m7922w6mun083kpjqfr9q9cal8zkqp0yaxyr4sy8v4re",
-  "bc1q9evu4ywa2tmhxt9x3fe6nt0e5ngxmq7kewmdvuk57j4c53ak4zjsj87pw7",
-  "bc1q9f5cecrwtwrfrvgxde5c04serf3txc8mdx7kgk3zhen2xp22zgcq5dngyh",
-  "bc1q9fftzadztvgpe90hnwec9n5al44xkuw6dc0n3mtt8w9wxyeg3rlqwfn527",
-  "bc1q9r4slky3fsjnen39u3fh6d22pwvk3pduj3htwyd5as3hywe2d49q6zyk05",
-  "bc1q9rrkgd0x5rv57ftysp3t4e35zhma2ejcjj0per8cg89um0mn0acs5hqh5k",
-  "bc1q9xne956dj0szmxur2up3xg4jl0ktjxpkggr7r5kw5tth48n4pemskcd9mc",
-  "bc1qa0mxdj0yehl3m99c8tsgepmngz0a0yf96v9al3ejyvla97hysjdsgzj59x",
-  "bc1qafrg9lt04d9wvurrgvr7l8p4cn5f07zh86d8rljeh6fzqegxe44sj82x4a",
-  "bc1qah2zmzd4atalqpds7ke86h7lvv49xqa43cyk28y2lwsus7sujd2sndplut",
-  "bc1qaj7w4a7kxvkjdgz76mc2uy66825ngjhqtasqd6tgssrajc68dk0qqkd3hc",
-  "bc1qak00ktehfya8j5vg5xcq7dfa2rae7zfnvjgs9jm0fsxa7a420rjs4zcp6t",
-  "bc1qanjru8xgu07kq6w0meyvydsfuvrlt75f4uekrhkuucp73w4gq0kql9dh26",
-  "bc1qatg2m52vj24zxvu92azvphxs08xp59prr86fcgdg42sp85v3qhsqsw88rc",
-  "bc1qau35ge0tgqlqvhn3pyv99ckukaxh9kes6e3jfpc4vw2ssfh3fpeq4dec6j",
-  "bc1qaygehcvv8334c6x9e42shrjz88kpnhw0gycz5fque4w4rjck6pnq7msl6g",
-  "bc1qaz28gc8e58uj5e80nu9j0yjws4zy84ahqlnfqd9ryhvc7a80vlyq430rvv",
-  "bc1qc058aal4cj2kk8e3tvhwu95cndy0lzr2zrxe49f3yasp693nyh7slfcgpl",
-  "bc1qc3h2hv8wnld70ujx8dyrmxs0ly27xs7us09y6e226yfl5vrgc2qsrxuucz",
-  "bc1qc9wlxjgu53ua5efqyjfwhwr3cn89smxq230h574v56edwmw0f0vq8qaakl",
-  "bc1qca7rh32jt8kg4kqtkcu98uucjlsslzsreh7tchk6hddvwe8kzzjshpppat",
-  "bc1qcf7wyjzz6f32je5vyk6pvketf5ykyg2lys66vwm43mc0g0gchzasy0k2e8",
-  "bc1qcmslz2gwsenf0xn9g90gvwkzympzxxhwhp8vrckudj0w5pdndwss92v3h4",
-  "bc1qcnp9057urvh2rrm09eya8mrasea7yys9q3q8hxs7unyct7nydtcsz57ajd",
-  "bc1qcnx924c7w3e8r42q3rvup60lw64gyn2kd4fs3pwckw7zw9966aaq568nz4",
-  "bc1qcudwp0xdlldgw78t6zfcklehwqt44a0hqt6n92k6uplk2w6q33fst0kdnv",
-  "bc1qcxzdm5pcfqnam3umsxsnwfykfx8j9rsl4nhrfqnt55qeg9q3625qvvyyuz",
-  "bc1qd3lvm6evz5suedhlwyrtf9cjrqtaq9unwe6423yl3j3exmq97a4s0mpzwt",
-  "bc1qdjxc3z0m2q5mz7lr67pmuq8hgqmhds6nmae0fkk2s07zge2f9jsseutev2",
-  "bc1qdm5rxs4w29c5u8wg7ppeqssgjrql23c0gsps7m0ctr6ratcl2a3sjnnq9g",
-  "bc1qdn29e8sa4xmwf0c4r6xkekea2mwt73d2qjsdtsne5j6gd9ujwysqmfh59v",
-  "bc1qdqs2zn6hrhvpxrtp0t82xfk2jv2j445hu7gguzm0ktwzdq5zzflqth7638",
-  "bc1qdvpvz88pq46qnvcz6jz08nmfldwhf8nx2v6x4z4tc79mpy8smkvqc9t9q8",
-  "bc1qe6xyfazpkry4nje0wmhfnv6faxfuu2drx4uryejdl2mhrplgeujs3knff6",
-  "bc1qe9q66vu5yyuzedfzugdh3yj5ask3n4wpx8pghw7ff6x7dd5p0gqqf66rxd",
-  "bc1qecswgzwg3sldnh4cwdfgly7f05sp2sela8g7hg0y0vp8l7t9nueqarz5w9",
-  "bc1qemj5cexuhn0ld93drv40kqfhvl2dag5vgy3d9vv4dar0hl6he3xs4fpjma",
-  "bc1qemtc7ddu4fvngf7pg8c5sukqswt3fr2zcnqnt7pju6a5fyrz7e5q96x25r",
-  "bc1qesf0u9lu5petyuukl9zq0g86wf2k69yehys2n4te7pf2y2csqpgst0tqhu",
-  "bc1qf285u7ygu9f2umg33a4c9wtdkfpahpt5kvqazk6arzktsw6vwguq02eckh",
-  "bc1qf2dh6acfagsayfprjzwzzysn4d4rdudkd6dq2adu2wj0t6ty8vlqnse73k",
-  "bc1qf3ret54m7mdevr8xwv8msstpcx6978vgwzfxkw3zjmlydfk4lqxqekert0",
-  "bc1qf42d25evlz8m503fgdw3trkzq9y53s5kwayfwjpvas26q2cnlt4skltleh",
-  "bc1qfdpu4yp2zk7xmm6tpfcatxe4gax5ktd0g9ra59jzcqtqkldrqe4qn0qqc2",
-  "bc1qfjrw5r0wttwu0u0gewyfcs2v37mujcxlhrfp7pqt48zwtjt2fs2shqjw2d",
-  "bc1qfrn20wgnt6l0a3qq4fxca8c66cuyqmvrfwkwwqhkxnf6z5xuxnpqzdzymf",
-  "bc1qg47q9u0ctw40jf0p8wswc67tcr96wel065f735cdu3e9aypqpumsvt4lun",
-  "bc1qg7egerrjv5ka0axjcwzh2rnkldr2k5yace828glhjyc6yz6s46sq6jk8xw",
-  "bc1qggrjslsem6ep9x4yxvyfdtgcxvyf9z85srlxpsuucrq6xsvaz3jspha83z",
-  "bc1qgp74f6xyajxznu2hnjh9txmzy7k7et3pkh9rc6enxt68fh8lw5js78s4d8",
-  "bc1qgpf2ar68l58zau2alfdnuw6afhm99k0nlfsrvcacd3jzpjlhq96qhxaxtr",
-  "bc1qgtldvmct0wa0569crgedr3ccrlydxfrymk7rfmalt7ehjqx54axsfv00gg",
-  "bc1qgtlgfqdyw60cawws4lvnesva46lgmyx2hhfl9xcnqp5m45u3dq3ska8nl2",
-  "bc1qgu5wpteehuuq8dcdd30d7la405j9r8wa2ttzqawt73lt8aec6kjs6u33r4",
-  "bc1qgzq2gq0l5nuxlaj45eyp4v3w6scaa9hx6f4chg60sufwmqwe4c7sp9tcng",
-  "bc1qh08szja77tucpmmwqmvyk72z0kjkqwuk2r327w7vwwh0f7997u6q3du9ga",
-  "bc1qh0yftewpsf8s3pyln8s6ky6qd6c7n643gndfhmkp7uczayua2mys9u6gw7",
-  "bc1qh3kulsckdz7d5ar3taglhqz4sa3wuasanyypuxyjscwzh0mm3g8qe9danw",
-  "bc1qh4qr0ytwjz7axxyrnl95la36jdqlygqpnul7wxx6a0mwzfmg9tkqkgszc7",
-  "bc1qhdaulu7pyhdhz4ty35v62gkx56nj2l7e6rz3x799sztdklan8n8sx863nu",
-  "bc1qhh2t6x7z57svrxdtl83xzdfz8melaueg7jmtpmsma40g2cfdxkes3lkd95",
-  "bc1qhjjcklsnjs0aseena8mmvlu7kufmeuhrlfz9l7dewpqsn43s707qkd7dny",
-  "bc1qhjxpm66t2k6v7c5zk6u2sdh8t36kas4a5402sqw82ewdpzuutf2s38ra4l",
-  "bc1qhl29w3rzcp9q3623hzvvnuksst0g48347qm82tr8ng3dfm9r0yvshlmvh2",
-  "bc1qhn7h9l2ky9l95sq7rqgaxp6c8ka0r5vfrghaln2pmeu67x4zpzzq3zcccf",
-  "bc1qhr5qw998er3rerm0lawlzzuuux6kja9zm4ryjls0aq5anprwuugs0yfcdg",
-  "bc1qhrgwkqetel4vpvj2azyyyhdvpg9u8j2dqz0svuwzh3vx4pvjf2dqtkt4uk",
-  "bc1qhrsvduw6j20dh4mseyf38zxqzd642zd6vmfjg9lmjj4d25qlvj5sv36ajm",
-  "bc1qhszm6vgtkhdn9cex8xurjkxjs2jhte0teetdnu4clr0ecxm20u5qhghyh9",
-  "bc1qj233g3tm9jejsnxvz3a06zqaqwwj2356f2n2sv572ggpahmmmxfqflfcx4",
-  "bc1qj390dgucrvx68jfad2wdpfhfuynvztgy9sy9a7009wmpecsdlu8q8zh0k4",
-  "bc1qj5dr2574mxgrxqxahymlvwna345d6d8424tyuqy3v2zhgu55j7fs2ezzq5",
-  "bc1qj9sat2haxfa7lsga96cxu26w9k4nr7z6sx4m33uzfq2x53sagcpqfzzjdz",
-  "bc1qjfmxg8fxqx286r254f3frzgra4sp4n5p2zczfsnrpkhdnyjyrhfqwul5wh",
-  "bc1qjnf3kfl592u85jlkvv99ms2qghv7yycrvg9902dtlh4lnfywg6wq5stn90",
-  "bc1qjnt2w6jx3ttx8dw0gt3nflm3ld0n2akkrkjzp73ua3kxzyhsj4jst848l6",
-  "bc1qjr7emlhxssap2npu5szrs5fa3cgxcnrd9gh9e3mdxzhdpye4puas7dp5pk",
-  "bc1qjsqyvnms83kvlnhc7r8phce0wj7j7nnvh8ye8ejxynke0uujwvksn2lely",
-  "bc1qjwjq6t60fwfsr7hf80yrhyy9lrvfjl2l9lpdwsw6sds4ph9uuglqjj53xx",
-  "bc1qk6ctvhc78eg80v989xmt7u5mwefewt728d9s0jxpgvqxa2m87xxsnu0wzw",
-  "bc1qk6n59st3mvj6ns0juv3vsdng48aszwdt7gyp0xlt8mrfn5qc9crsdx8dv8",
-  "bc1qkayh60m2uc86cere4z2gz03g2ayg2k726klqf0vx7lq588xzz0nqjvrg0q",
-  "bc1qkeh8sh6hzc49097j6fs8lgkfxpyas7wntelwnfqk4xhzccaj0nfq73s3f0",
-  "bc1qkjr6ew6d5nqf2vpxd7vm4w5lps72c8sxnc7hxnwcs9psqvpcmqkset3kdn",
-  "bc1qkka8ku2z79wq7wtkr44qmp6x3gftc9suju2q3ncw0xdy2nwj37jqu7hec8",
-  "bc1qkkhxgzvlvhymm2eprt0eukr3gw23zmarcuw087td9l34tmuprrysh643xt",
-  "bc1qkqap38eq4y8auh6z8yjwrj8r2p77n6xshcp8xpy2kkjcu7lhfyfqntsrjh",
-  "bc1qkrh06leyzgp84z0v495c0gr68tqn9ewue3lendpcrdz8htz0yw6q9aqyz3",
-  "bc1qktk40875p4gcafun6m2yzfxsrdt6d5h55c2zhd45wfefkusux27s38c764",
-  "bc1qku8s2753ttheldjrz5v4xmampnetjp3vy30m7j0nd3348e2m9rpq26l3ef",
-  "bc1qkvtt6kaw688mzu7txj9q3mn9m60pskv62gdz5j5kugd5fc7f8dfsdsfsq6",
-  "bc1qlftdm0w434ud5nz5ktnmak4gmr7k4x5qajmr5pvuh4vz9m4e8ljsujemsu",
-  "bc1qlk5g7mtuyxdy3txqz7q24yjffcdk6g6mr4ks48afv5234hly3hvslw8aje",
-  "bc1qllpem4drxvxu45r7aj7aqma50cylmaup07eynehs6v3f8zppcw2s00v3rt",
-  "bc1qln0rhtex8dwmqgex2mzkm47c437tunl6kw9amq7s4dye57fjk44s8x264n",
-  "bc1qlss2xnh3hsa78pedunw9qsm9nzpczrq90dtcp0sqknt3xkf5wyzsjtc662",
-  "bc1qlv8j6wgysv9xdpqtfau2lq5rcju0v4eyazrlawpgu8pwd2usz0dsfqc9up",
-  "bc1qlv8t6axsf50pe5wnt4x4xhqfudkumdvx3rc38f9kj3q82qhqy6lq626v47",
-  "bc1qm6hw09d24eqlz0x5r5txlpsrqvpf250fd5py3s4uwax6gwtdjfrq03yh9z",
-  "bc1qm72da6c6sq6a34rkwxk9sfpmy9kqea4dv33280hm5ghfdm4myvaqxnhcrw",
-  "bc1qm8z7nqmtmrurzcz82yec3sgk5rz67tm5rcaklqxajyxd348vppes86jvuz",
-  "bc1qmcvwrhj556capug3q5xzhwyhq82cps4l6ng0w573v6rl8zgkrlmqwh9386",
-  "bc1qmfzk56r5xvgfztej8fkg7rw76wy5qylh79shsaewpltfq4pz2qss5ajj7r",
-  "bc1qmhjm7sca3qwyujgfqtg3vw7swzxyg7mkxe4e7uzdef5ftc2kejwqjp7q7p",
-  "bc1qmj4g65rdqztaea7aqwnht5v2fhajmjat8j0q75nencrq2gu8q3ds5e0k5l",
-  "bc1qmr28gws3f589wmxf870mee4mapyzrv2209h3c992hhv3nn7ppctqhejqve",
-  "bc1qmschx4zu5yd9y82jqs6rwse3lftuzu97j9x76wdtx2avtee0g56qppefym",
-  "bc1qn3r3kqf90txgcsdxzu2s0p5avp79axlf98lzrwlmpnmxc3ak0vwsuxrepa",
-  "bc1qn55zpxs6fzjju36zuvkndzjfz0glhejtv299krn96sahl2pupthqhjc49k",
-  "bc1qneexyhzwx2g7ugw7m93kf48d6mdpt76gghmnr07s7dq7nhlwvc0shnryqd",
-  "bc1qnnx8uu6tuw5qlnqava2vjzrlsrwlaws7h6med4ndhtghrw8r5vcske26rv",
-  "bc1qntgp8cuk9sp2xadzc72h20qt372nvkjk5pmknvu9zntnf0yly5dqy7ev8j",
-  "bc1qnwa48yav84vzadrk79ulzrzu8s5epl7sfqelxufnpls63j7szx6s4n20hg",
-  "bc1qnwh5kqa6u5k2zevq0dfwn8z8mux4sz6t23g4xqeaf2lrd6zda75ssrs0gx",
-  "bc1qnx38lf4ds0lrw3s0l9fc745l0x3xdzw3w8stmxdvwu3v0n8jrcqsxm34ls",
-  "bc1qp798q5d3n8wsj3x345c7v65u9fwfe43e9vyupllmrrjq5k5n39rseze5vk",
-  "bc1qpa3ggjgfsr4cule3c42ns4y97k9z38whpx5940uw4mn6taeaaqjs97t28r",
-  "bc1qpcywcjgfel39zc762pmwayj9wu3v5xly0kjtdx2n2w0pandsce9q5rr54q",
-  "bc1qpr7fnnc9gan8vd9waym2rc2td24s72732wltnu9rc0yxhdssqaasr3up6s",
-  "bc1qpsftpydc86wxgga26r6nfq5tvrh5m34055kdrp25lulqu70shunskxflr3",
-  "bc1qptacp904tazvu54qxty5qunnwqhcm84ydrt39r9c4cqzkqrtju6qfd7eyt",
-  "bc1qpvwcac0hah03lepsvrx8zu9gdyvprfnckga4cfjg3ctaflwc4xvqp6hw3h",
-  "bc1qpwhmpvrau7f82hlh6es4wzedcq2rx5aqx5ue4ajsl4r7h29udx7q2ztulw",
-  "bc1qq5d4g9gcrwmwm8w965cwtkdgu60sl7fteahhtelzhna7rm29q0uswgca9t",
-  "bc1qq6jmww5tpr6t4cgl577d28cjr6mpxz85d6tkh7kjn2njkyd7flrqyt25yw",
-  "bc1qq77frdehk027ur54wzhhpg8dtqwpdn3emaxv42kxxlzmttrznzqs7mye34",
-  "bc1qq7m25904sfa4qn0na5gfssltjdmxwuvqxfl73dunza4s4cysqh7ssale4g",
-  "bc1qqea6940qjuhnuqqcu6u5ajeh3as8drwwwey56vxp90d05gzx72psw4ngua",
-  "bc1qqf4k7r66y3g3fpjlq6g46g5av5qgqvjpdhu6lhg68fwzm4rc4v0quux3vy",
-  "bc1qqfkncsw0gzr3vdyearnvt9w3ckuymet0v495dk5hwpfyjfeej42s28g3uz",
-  "bc1qqmsz5fe3dqcg7wxt3wc9s69td0yxnhlj97pvxknv8s25yuylawwsc99267",
-  "bc1qqsse9hgmx3w56c5suhkp2gqyxqhh6acc6wf0gnd7fjtp97f7c8esmhtj4x",
-  "bc1qqtmu9k5pakw4q4p242322x5tjy6tqtjj3u596nvksxjdhw789gfsmuskwz",
-  "bc1qqtyxkv77ce6s7hqw2ljg6axaxf4463q2krwne5vlxhhk3x7laasq3vqe5x",
-  "bc1qqx37qn2ca3gpjtuj5dkh8l6u4dh9sujn66f5gwe5maw7cvw47rmsej5kwl",
-  "bc1qqzq92mhj6a0v8kflqx6t0p3g47pdrqpv8f8gu7uk2a30kxxnplns4ez65a",
-  "bc1qr9kjw52l7m3nxhprgkt99nhrqtem5erzay6jy3j7xzyge65z3v2qse5xdw",
-  "bc1qra0g63370unwjrjpx4f267xss94zgxp78wqyc8gavqstnplakmfq0yq9f9",
-  "bc1qrg5y53hvpnu78pvy70v4s0qe52rnx3298gdvc7t9lsu2lya7deuql4a28j",
-  "bc1qrje4nstr7yy2qvmrfazlnjcsd3573fh6j9u58t446mn3ltp8e5vqgz8535",
-  "bc1qrjgp0ksa0wmhdfslzta0cem3s030jg3cxf7xfqtjcjy3jstj8wnsenhgs3",
-  "bc1qrmhgzejkazuezzduws22wuk6atlsxruytyh9ynx2855qpfezy5vskdznmv",
-  "bc1qrpcnmqxe8q4mq3n04qdy657wu23qxu8x5pvc0yx9adwz0x093gtsh5r2cv",
-  "bc1qrtyl2hpxu4lpl8x7fz6rlwvd43v6a9ynvqjf2mr7gegvnmdc0lxqas83mx",
-  "bc1qruqm8ef6lukm3kew536vxpys0nsrkdqqq7gn5eyjz6fsdee55frq9ksm66",
-  "bc1qrz3dctwmkkksya4n0cfhfdej480vx9shkt9vfe63qg0crmr3x68swt9jn8",
-  "bc1qs3w6eh4wqxphz769jlkglhdrmfaz4pj668aag2qrhcmrj74a02psr7mx0l",
-  "bc1qs6edu8cf7dtn2fq4earyhagryd3zsqvkdlaj285xptzjrq4szzkqz3fxze",
-  "bc1qsh9f8ackh9etwqjy6hd3vcqnc8wy9wnynsfypyv0p8nxesytkqeqtp7edq",
-  "bc1qsl92pxqza6xmw26xp6783slr46eh7wfdeyxnpyg0ammfcrc370esjrcfma",
-  "bc1qsmwlvh7v6wkg9kxtgh65q9lat00xx68g7nqycn8jdr5dudvqe8ksd4mxg7",
-  "bc1qsu0lfh6ae4pklru5cs6qscf9z88m4ckljlcxyqp5qfcvv5ygg40suzkygv",
-  "bc1qsvgk35j2tgq3anjydqlccunngrqur9pk3hdlfpgmd3kq69njaqmqdde6r0",
-  "bc1qswcfs3404reuplxzvmnxv23r82jzttw7ugm4gzwdwddpqa985mzs4ex3wv",
-  "bc1qsy0tanmqylsxjzpfecms96du4vp4tursfh9fyp4w0ff5aky7qe9sclj6f4",
-  "bc1qt5fugtlsp402yta0xvl79rlxc5epx9nd926f9dvct5r3uzuay5gqwyq55w",
-  "bc1qt5hqtxfktkuusrpcyus2zrc4x28hg23l35pushfmj0dsvmfrv2lq5vu9au",
-  "bc1qt8wuxew7ly35fvu2xaytmjvr6dlux5d05p9qahzw6h3axkc5p2ksla52xp",
-  "bc1qtahzsvs0hrt22dmdt3268hpdh9eh3rlxprac584x8l9s2df6jefq5nse9k",
-  "bc1qtcx3lve9uscxevqv0ekj2e75kjy6mqn0aedxq6xrtxv9td0jm7dsj7vtuu",
-  "bc1qthgusw4jxlw2e9h69r8hpv22df47dups2gmkzlgp7l8ww27eew6q7ewjvv",
-  "bc1qtjq2ypqq3kamya3dsfphy7mdqwp5nc4adaxan6mqdze59d7ptu3qc9ryth",
-  "bc1qtllzgt5hsxvnhga8tq3nars4d94c263q8dm8vakrq6r42648vhtqtyy36q",
-  "bc1qtr7kzsz669mfn8677ltxlm8c3t9nkzskj2v3lexx0kxhamzg7zxspkd8w9",
-  "bc1qttkf88qadw9mlvnfftw37gllk8q7m3jmfxrugath75rl3ydpt9zqlfjzzl",
-  "bc1qtuvqgvkjnq0gw4cc6p60yc3y6j7jzz937cs0wy6w09h5dyg36d0q8xsjlh",
-  "bc1qty4wtp9zhvtjuj7p36vf4avs98l4vata8634m3z4ys99t5qlhjwscslcg5",
-  "bc1qtyaqju2ar9x2v6c4jf42j9x9qzjcwev9u8uys524qnhznths37zq27ppfm",
-  "bc1qtym0jxefs7z9vr5y50gdsmtdfgcfyr9rxema6n23jw4cswulcu6qsj8t92",
-  "bc1qtzn5ktk3hhhgaweupzt9k6k9m00jws3xn5teplwupfy2ltl0xcjsl4g02l",
-  "bc1qu5pttfk0s407dtr7ds2e547ugq80ay4n82qpvwqzsvlccdjcp84qxpef24",
-  "bc1qu6jn53gw9t3j62lee0ht0p3dhtkycg24gj5kcp230jq6fn4jcsnsxjxn6h",
-  "bc1quezt5y4lqf37guytyf8jj56eywjvsanv7u8fq6q8p46hg02tc5vsyx3let",
-  "bc1qup8qwemtepa2wf4r8w599lphxq6f38s3ta8925eu7y8zvp4x6v2svqrfq2",
-  "bc1qv85ky08r24v3cnysh8hk9c5ftg60f2kwellmzpzlqnwfawgtdqesghlznz",
-  "bc1qvax0qf0zgycnwyd2cmgzsr65l6mldc7r3wm4876xg4u7rtywp2dsw034tt",
-  "bc1qvcwzkv448c9rz60fas5wqgnrl5gzvh0nnacyudqcsxhr0j9fa24s48m79n",
-  "bc1qvkn8eymy7c2rpw2c5x6x8td4hrul6vx79y4mtd4zzdxdrxwk9q2sv6rnn8",
-  "bc1qvu6u5tnjk22sggwkp2j3vlxle5mq9xz8v82g7j57k9ele02wjmkqz9vrvt",
-  "bc1qw0lsnwv0sga0n3rcpaxtjg4z6jhf0xu6dyhdv7jcx9d8gm4phtesr9vywe",
-  "bc1qw5ue35s2rms9wz2ajxu4h83uzaylax7jqwp3h3cxamj9j05dgw9q6nnr24",
-  "bc1qwu39fckqy2cl9z42994xyhg657q7txajwakf9sdzt4f3caluk3fqz3zl94",
-  "bc1qwx5sl8mls6hkmumjnekf6szuzh979cmgzsz8dqm3xeayrxfgz4us8m8y4z",
-  "bc1qwzy2ezy8zskmuu2v6y60w5uqe46qeql94may8fetnl9uyahlx0lsthg04q",
-  "bc1qx628pacq3m5px4zjjagrtpc8a6v3my5l06rxcrl5xr3vfv4s00cqcxmlhy",
-  "bc1qx6ws8f6p3zzevs0rf5kvu7e074snfmr8mms8ezgreg7kas84l7ssptyejn",
-  "bc1qx9yj7x4xmek7rm5styss3vcdth48p6uqzjdfsua9s7v430jq2r8q0j4kge",
-  "bc1qxaj93j90njr4lwrevjhg3zr2qyr0zerw0d8fxlxhep6c8uu4jqdqhlqptt",
-  "bc1qxpa34lagwsvq5t8ul09y2hz2dh0flkampx8lqyp3jdt4lpwaxx2qe2uvxj",
-  "bc1qxs6nw2cjmctjeqjk5lhr47uv6anmge3ut66gmuexcryrymusqt9s2tvz62",
-  "bc1qxuv7ge2hcfhzykp49g2pxz69p99d0aujx5wupud28tulftlc0agslcmuuf",
-  "bc1qxyx06rxxprx9plp009nvl0f0ft5e3af3tt53n7pctylafqfltwvqyk7sfg",
-  "bc1qy2av32x3tkfucu25dqfh4nqkf50ayng43ervl7fcwg4crcu3queqtlcrsa",
-  "bc1qy7efv6fkened6hnhzydmcd9eatj20dsnc4px2gtmkedc0xznlcxq6n04jv",
-  "bc1qy8hl8khkpcsnjqvgna6h8249x4at9vjern9meph0e5el4g0nudlq7kd773",
-  "bc1qy9f5q32h9rmem40nd06e4l8cxwqlttuale9s6q076aymt0dayx6qekhfmp",
-  "bc1qy9w004s6gek29k3npgmxq8huqx48l2gatx9yqlu0rchxzlnn7nlsn059c2",
-  "bc1qyr5df2mhp3lfjrd2sjx56eqm7netrtdtwengw657k7k6hnmyedlqwfx472",
-  "bc1qys46dh5q94lykgqrvthhh52ajsk7s4c5ex7p9t7tvqe3uxheq2us3rwra2",
-  "bc1qywh2j5t52zhqhvlqymsw78mq9lfrh4utf47ufhfpkgfzj3trqaaswuak4h",
-  "bc1qz2fe23n5ykggw9394gxly6jqxk23rmhxesvd7q9d3sfee6cqmycqha09qg",
-  "bc1qzchnul757l7q2yrhtv2gga3h560vfjdj002xcstcd5cvspa3puks4h7gu5",
-  "bc1qzdah777dzr46n8wwdmr2y8etxzt0e66j90v4f9m2dj2wzjk3mtgqhamaj5",
-  "bc1qze2smuhn3k3yxq49fzsmz4pfjemnhpsjt5g3xc8slyh42edy23zsxjsr73",
-  "bc1qzk3kkegykpk0uzyekqfyr4l4ky62hx7tt9cx4r5klv98x3f37z0sr26gaf",
-  "bc1qzlhsk5ftpz6uaxgmm8hel59v7avmn0q8wp9gaaxdzzkde7ryumeqdnte02",
-  "bc1qzm3t8m5fcnz8t22kf2lyr97u2wret0pkjvzngldnqq5su434tsrqxjncvj",
-];
-
-async function tvl(api){
-  if(api.timestamp > Date.now()/1e3 - 3600){
-    return sumTokensExport({ owners })(api)
-  } else {
-    return {
-      [ADDRESSES.ethereum.WBTC]: (await sdk.api.erc20.totalSupply({ target: ADDRESSES.ethereum.WBTC, block:api.block })).output
-    }
-  }
-}
-
-module.exports = {
-  bitcoin: {
-    tvl,
-  },
-  methodology: `TVL for WBTC consists of the BTC deposits in custody that were used to mint WBTC`,
-};
